requirements.txt
requests>=2.28.0
google-genai>=0.4.0
google-generativeai>=0.4.0
notion-client>=2.0.0
psutil>=5.9.0
